chore: move mappers to where they are used
